Split body_writer into its constituent parts
No other changes are made.
